RET-2843: Updates after PR feedback
