+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/CaseActionsForCaseWorkerController.java
@@ -26,7 +26,7 @@ import uk.gov.hmcts.ethos.replacement.docmosis.helpers.BFHelper;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.FlagsImageHelper;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.HearingsHelper;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.Helper;
-import uk.gov.hmcts.ethos.replacement.docmosis.helpers.RespondentService;
+import uk.gov.hmcts.ethos.replacement.docmosis.helpers.NocRespondentHelper;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.dynamiclists.DynamicDepositOrder;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.dynamiclists.DynamicJudgements;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.dynamiclists.DynamicRespondentRepresentative;
@@ -46,7 +46,7 @@ import uk.gov.hmcts.ethos.replacement.docmosis.service.EventValidationService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.FileLocationSelectionService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.FixCaseApiService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.JudgmentValidationService;
-import uk.gov.hmcts.ethos.replacement.docmosis.service.RespondentRepresentativeService;
+import uk.gov.hmcts.ethos.replacement.docmosis.service.NocRespondentRepresentativeService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.ScotlandFileLocationSelectionService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.SingleCaseMultipleMidEventValidationService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.SingleReferenceService;
@@ -106,9 +106,9 @@ public class CaseActionsForCaseWorkerController {
     private final DepositOrderValidationService depositOrderValidationService;
     private final JudgmentValidationService judgmentValidationService;
     private final Et1VettingService et1VettingService;
-    private final RespondentRepresentativeService respondentRepresentativeService;
+    private final NocRespondentRepresentativeService nocRespondentRepresentativeService;
 
-    private final RespondentService respondentService;
+    private final NocRespondentHelper nocRespondentHelper;
 
     @PostMapping(value = "/createCase", consumes = APPLICATION_JSON_VALUE)
     @Operation(summary = "create a case for a caseWorker.")
@@ -278,7 +278,7 @@ public class CaseActionsForCaseWorkerController {
                     && caseData.getEcmCaseType().equals(MULTIPLE_CASE_TYPE) ? YES : NO);
 
             //create NOC answers section
-            caseData = respondentRepresentativeService.prepopulateOrgPolicyAndNoc(caseData);
+            caseData = nocRespondentRepresentativeService.prepopulateOrgPolicyAndNoc(caseData);
         }
 
         log.info("PostDefaultValues for case: {} {}", ccdRequest.getCaseDetails().getCaseTypeId(),
@@ -427,9 +427,9 @@ public class CaseActionsForCaseWorkerController {
         if (errors.isEmpty()) {
             //Needed to keep the respondent names in the rep collection sync
             if (!isEmpty(caseData.getRepCollection())) {
-                respondentService.amendRespondentNameRepresentativeNames(caseData);
+                nocRespondentHelper.amendRespondentNameRepresentativeNames(caseData);
             }
-            caseData = respondentRepresentativeService.prepopulateOrgPolicyAndNoc(caseData);
+            caseData = nocRespondentRepresentativeService.prepopulateOrgPolicyAndNoc(caseData);
         }
 
         log.info(EVENT_FIELDS_VALIDATION + errors);
@@ -464,7 +464,7 @@ public class CaseActionsForCaseWorkerController {
         if (errors.isEmpty()) {
             //add org policy and NOC elements
             caseData.setRepCollection(updateWithRespondentIds(caseData));
-            caseData = respondentRepresentativeService.prepopulateOrgPolicyAndNoc(caseData);
+            caseData = nocRespondentRepresentativeService.prepopulateOrgPolicyAndNoc(caseData);
         }
 
         log.info(EVENT_FIELDS_VALIDATION + errors);
@@ -483,7 +483,7 @@ public class CaseActionsForCaseWorkerController {
         log.info("AMEND RESPONDENT REPRESENTATIVE SUBMITTED ---> "
             + LOG_MESSAGE + callbackRequest.getCaseDetails().getCaseId());
         try {
-            respondentRepresentativeService.updateRepresentativesAccess(callbackRequest);
+            nocRespondentRepresentativeService.updateRepresentativesAccess(callbackRequest);
         } catch (IOException e) {
             log.error("Failed to update respondent representatives accesses");
             throw new RuntimeException(e);
