+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/CcdCaseAssignment.java
@@ -1,13 +1,5 @@
 package uk.gov.hmcts.ethos.replacement.docmosis.service;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
@@ -29,8 +21,6 @@ import static java.util.Objects.requireNonNull;
 public class CcdCaseAssignment {
 
     private static final String SERVICE_AUTHORIZATION = "ServiceAuthorization";
-
-    public static final String BEARER = "Bearer";
 
     private final RestTemplate restTemplate;
 
@@ -67,22 +57,6 @@ public class CcdCaseAssignment {
                 callback,
                 createHeaders(serviceAuthorizationToken, userToken)
             );
-
-        log.info("authToken: {}", userToken);
-
-        ObjectMapper mapper = new ObjectMapper()
-            .registerModule(new Jdk8Module())
-            .registerModule(new JavaTimeModule())
-            .registerModule(new ParameterNamesModule(JsonCreator.Mode.PROPERTIES))
-            .registerModule(new JavaTimeModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-            .setSerializationInclusion(JsonInclude.Include.NON_EMPTY)
-            .disable(SerializationFeature.FAIL_ON_EMPTY_BEANS);
-
-        try {
-            log.info(mapper.writeValueAsString(requestEntity));
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
 
         ResponseEntity<CCDCallbackResponse> response;
         try {
