+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/service/NocNotificationServiceTest.java
@@ -11,7 +11,7 @@ import uk.gov.hmcts.et.common.model.ccd.CaseData;
 import uk.gov.hmcts.et.common.model.ccd.CaseDetails;
 import uk.gov.hmcts.et.common.model.ccd.types.Organisation;
 import uk.gov.hmcts.et.common.model.ccd.types.RespondentSumType;
-import uk.gov.hmcts.ethos.replacement.docmosis.helpers.RespondentService;
+import uk.gov.hmcts.ethos.replacement.docmosis.helpers.NocRespondentHelper;
 import uk.gov.hmcts.ethos.replacement.docmosis.utils.CaseDataBuilder;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -34,7 +34,7 @@ class NocNotificationServiceTest {
     private CallbackRequest callbackRequest;
 
     @Mock
-    private RespondentService respondentService;
+    private NocRespondentHelper nocRespondentHelper;
 
     @BeforeEach
     void setUp() {
@@ -109,7 +109,7 @@ class NocNotificationServiceTest {
         RespondentSumType respondentSumType = new RespondentSumType();
         respondentSumType.setRespondentName("Respondent");
         respondentSumType.setRespondentEmail("res@rep.com");
-        when(respondentService.getRespondent(any(), any())).thenReturn(respondentSumType);
+        when(nocRespondentHelper.getRespondent(any(), any())).thenReturn(respondentSumType);
         nocNotificationService.sendNotificationOfChangeEmails(callbackRequest, caseData);
         // Claimant
         verify(emailService, times(1)).sendEmail(any(), eq("claimant@represented.com"), any());
@@ -137,7 +137,7 @@ class NocNotificationServiceTest {
         respondentSumType.setRespondentName("Respondent");
         respondentSumType.setRespondentEmail(null);
 
-        when(respondentService.getRespondent(any(), any())).thenReturn(respondentSumType);
+        when(nocRespondentHelper.getRespondent(any(), any())).thenReturn(respondentSumType);
 
         nocNotificationService.sendNotificationOfChangeEmails(callbackRequest, caseData);
         verify(emailService, times(0)).sendEmail(any(), any(), any());
